+++ NEWS.md
@@ -1,7 +1,7 @@
 # ggblanket 5.1.0900
 
 * Removed some unnecessary messages.
-* Changed default continuous `pal` to mako from viridisLite.
+* Changed default continuous `pal` to `viridisLite::mako`.
 * Fixed bug where `col_breaks` was not working for a numeric `col`.
 * Fixed bug where `str_keep_seq` was not working for date, datetime or time class.
 * Fixed bug so that time class variables work.